+++ src/components/App.js
@@ -17,7 +17,7 @@ import EventHandler from "./EventHandler.js";
 const App = () => {
   const { state, dispatch } = useStore();
 
-  const onPaste = () => handleLinkPaste(event, dispatch);
+  const onPaste = () => handleLinkPaste(event, state.dialogOpen, dispatch);
   const onKeyDown = (event) => handleKeydown(event, state.dialogOpen, dispatch);
 
   useEffect(() => {
