+++ src/components/ui/IconButtons/Button.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const StyledButton = styled.button`
+  background: none;
+  font-family: "Segoe UI";
+  outline: none;
+  cursor: pointer;
+`;
+
+export default StyledButton;
